# back/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the API documentation for Easy DEVS.",
		"title": "Easy DEVS API",
		"contact": {},
		"version": "1.0"
	},
	"host": "localhost:3000",
	"basePath": "/",
	"paths": {
		"/ai/generate-diagram": {
			"post": {
				"security": [
					{
						"BearerAuth": []
					}
				],
				"description": "Sends a prompt to OpenAI to generate a diagram in JSON format based on a strict schema.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["AI"],
				"summary": "Generate a diagram",
				"parameters": [
					{
						"description": "Data required to generate a diagram",
						"name": "body",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/handler.GenerateDiagramRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Generated diagram",
						"schema": {
							"$ref": "#/definitions/response.DiagramResponse"
						}
					},
					"400": {
						"description": "Invalid request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "AI processing error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/ai/generate-model": {
			"post": {
				"security": [
					{
						"BearerAuth": []
					}
				],
				"description": "Sends a prompt to OpenAI to generate a DEVS model code.",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["AI"],
				"summary": "Generate a model",
				"parameters": [
					{
						"description": "Data required to generate a model",
						"name": "body",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/handler.GenerateModelRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Generated model code",
						"schema": {
							"$ref": "#/definitions/response.GeneratedModelResponse"
						}
					},
					"400": {
						"description": "Invalid request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"500": {
						"description": "AI processing error",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"handler.GenerateDiagramRequest": {
			"type": "object",
			"properties": {
				"diagramName": {
					"type": "string",
					"example": "MyDiagram"
				},
				"userPrompt": {
					"type": "string",
					"example": "Create a software architecture diagram"
				}
			}
		},
		"handler.GenerateModelRequest": {
			"type": "object",
			"properties": {
				"modelName": {
					"type": "string",
					"example": "MyModel"
				},
				"modelType": {
					"type": "string",
					"example": "DEVS"
				},
				"previousModelsCode": {
					"type": "string",
					"example": "Existing model code"
				},
				"userPrompt": {
					"type": "string",
					"example": "Generate a model based on the previous code"
				}
			}
		},
...
		"response.Connection": {
			"type": "object",
			"properties": {
				"from": {
					"description": "obligatoire",
					"allOf": [
						{
							"$ref": "#/definitions/response.Endpoint"
						}
					]
				},
				"to": {
					"description": "obligatoire",
					"allOf": [
						{
							"$ref": "#/definitions/response.Endpoint"
						}
					]
				}
			}
		},
		"response.DiagramResponse": {
			"type": "object",
			"properties": {
				"connections": {
					"description": "obligatoire",
					"type": "array",
					"items": {
						"$ref": "#/definitions/response.Connection"
					}
				},
				"models": {
					"description": "obligatoire",
					"type": "array",
					"items": {
						"$ref": "#/definitions/response.Model"
					}
				}
			}
		},
		"response.Endpoint": {
			"type": "object",
			"properties": {
				"model": {
					"description": "obligatoire",
					"type": "string"
				},
				"port": {
					"description": "obligatoire",
					"type": "string"
				}
			}
		},
		"response.GeneratedModelResponse": {
			"type": "object",
			"properties": {
				"code": {
					"type": "string"
				}
			}
		},
...
		"response.Model": {
			"type": "object",
			"properties": {
				"components": {
					"description": "optionnel",
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"id": {
					"description": "obligatoire",
					"type": "string"
				},
				"ports": {
					"description": "optionnel",
					"allOf": [
						{
							"$ref": "#/definitions/response.Ports"
						}
					]
				},
				"type": {
					"description": "enum obligatoire",
					"allOf": [
						{
							"$ref": "#/definitions/response.ModelType"
						}
					]
				}
			}
		},
		"response.ModelType": {
			"type": "string",
			"enum": ["atomic", "coupled"],
			"x-enum-varnames": ["ModelTypeAtomic", "ModelTypeCoupled"]
		},
		"response.Ports": {
			"type": "object",
			"properties": {
				"in": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"out": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
...
	}
}